% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACVr-package.R
\docType{package}
\name{PACVr-package}
\alias{PACVr}
\alias{PACVr-package}
\title{PACVr: Plastome Assembly Coverage Visualization}
\description{
Visualizes the coverage depth of a complete plastid genome as well as the equality of its inverted repeat regions in relation to the circular, quadripartite genome structure and the location of individual genes. For more information, please see Gruenstaeudl and Jenke (2020) \doi{10.1186/s12859-020-3475-0}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michaelgruenstaeudl/PACVr}
  \item Report bugs at \url{https://github.com/michaelgruenstaeudl/PACVr/issues}
}

}
\author{
\strong{Maintainer}: Michael Gruenstaeudl \email{m_gruenstaeudl@fhsu.edu} (\href{https://orcid.org/0000-0002-1666-1773}{ORCID})

Other contributors:
\itemize{
  \item Gregory Smith [contributor]
  \item Nils Jenke [contributor]
}

}
\keyword{internal}
